+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/event/data/EnrollmentTimeFieldSqlRenderer.java
@@ -28,7 +28,6 @@
 package org.hisp.dhis.analytics.event.data;
 
 import static org.hisp.dhis.analytics.AnalyticsConstants.ANALYTICS_TBL_ALIAS;
-import static org.hisp.dhis.analytics.util.AnalyticsSqlUtils.quoteAlias;
 import static org.hisp.dhis.common.DimensionalObject.PERIOD_DIM_ID;
 import static org.hisp.dhis.commons.util.TextUtils.EMPTY;
 import static org.hisp.dhis.util.DateUtils.plusOneDay;
@@ -49,6 +48,7 @@ import org.hisp.dhis.analytics.EventOutputType;
 import org.hisp.dhis.analytics.TimeField;
 import org.hisp.dhis.analytics.event.EventQueryParams;
 import org.hisp.dhis.common.DimensionalItemObject;
+import org.hisp.dhis.db.sql.SqlBuilder;
 import org.hisp.dhis.period.Period;
 import org.hisp.dhis.program.AnalyticsPeriodBoundary;
 import org.hisp.dhis.program.ProgramIndicator;
@@ -58,6 +58,9 @@ import org.springframework.util.Assert;
 @Component
 @RequiredArgsConstructor
 class EnrollmentTimeFieldSqlRenderer extends TimeFieldSqlRenderer {
+
+  private SqlBuilder sqlBuilder;
+
   @Getter private final Set<TimeField> allowedTimeFields = Set.of(TimeField.LAST_UPDATED);
 
   @Override
@@ -159,7 +162,7 @@ class EnrollmentTimeFieldSqlRenderer extends TimeFieldSqlRenderer {
   }
 
   private String toSqlCondition(Period period, TimeField timeField) {
-    String timeCol = quoteAlias(timeField.getField());
+    String timeCol = sqlBuilder.quoteAx(timeField.getField());
     return "( "
         + timeCol
         + " >= '"
